Page 폴더와 폴더 하위 구조 정리
