+++ client/src/components/ResetPassword.tsx
@@ -1,7 +1,7 @@
 import { styled } from "styled-components";
 import { useState } from "react";
 import { Link } from "react-router-dom";
-import { useToast } from "./hooks/useToast";
+import { useToast } from "../hooks/useToast";
 
 function ResetPassword() {
   const [email, setEmail] = useState("");
