+++ client/src/hooks/useToast.ts
@@ -1,4 +1,4 @@
-import useToastStore from "../../store/toastStore";
+import useToastStore from "../store/toastStore";
 
 export const useToast = () => {
   const showToast = useToastStore((state) => state.addToast);
