+++ client/src/pages/EventDetail.tsx
