+++ client/src/pages/RecruitForm.tsx
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import { useForm } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
 import { useMeetingStore } from "../store/meetingStore";
-import Button from "./common/Button";
+import Button from "../components/common/Button";
 
 interface FormInputs {
   exerciseType: string;
